requirements.txt
PyQt6>=6.4.0
requests>=2.28.0
beautifulsoup4>=4.11.0
opencv-python>=4.7.0
numpy>=1.23.0
Pillow>=9.3.0
pyzbar>=0.1.9
cryptography>=38.0.0
python-dotenv>=0.19.0
pyOpenSSL>=22.0.0
SQLAlchemy>=1.4.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0 